+++ schema.yml
@@ -191,6 +191,44 @@ paths:
           $ref: '#/components/responses/Unauthorized'
       tags:
         - dissertation_location
+  /adviser:
+    get:
+      description: "Return advisers list"
+      operationId: advisers_list
+      parameters:
+        - in: query
+          name: limit
+          schema:
+            description: Number of results to return per page.
+            type: integer
+        - in: query
+          name: offset
+          schema:
+            description: The initial index from which to return the results.
+            type: integer
+        - in: query
+          name: search
+          schema:
+            description: A search term which will look up at name
+            type: string
+        - $ref: '#/components/parameters/Accept-Language'
+        - $ref: '#/components/parameters/X-User-FirstName'
+        - $ref: '#/components/parameters/X-User-LastName'
+        - $ref: '#/components/parameters/X-User-Email'
+        - $ref: '#/components/parameters/X-User-GlobalID'
+      responses:
+        '200':
+          description: OK
+          content:
+            application/json:
+              schema:
+                $ref: '#/components/schemas/AdviserList'
+        '400':
+          $ref: '#/components/responses/BadRequest'
+        '401':
+          $ref: '#/components/responses/Unauthorized'
+      tags:
+        - adviser
   /dissertations/{uuid}/:
     get:
       description: "Return dissertation's detail of the user currently connected"
@@ -1071,6 +1109,32 @@ components:
           type: array
           items:
             $ref: '#/components/schemas/DissertationLocation'
+    AdviserList:
+      type: object
+      properties:
+        count:
+          type: integer
+        next:
+          type: string
+          format: uri
+          nullable: true
+        previous:
+          type: string
+          format: uri
+          nullable: true
+        results:
+          type: array
+          items:
+            $ref: '#/components/schemas/Adviser'
+    Adviser:
+      type: object
+      properties:
+        uuid:
+          type: string
+          example: "f8b61488-f32b-4382-8633-3bf81fcacatb"
+        name:
+          type: string
+          example: "Eric Judor"
     AcceptedLanguageEnum:
       type: string
       enum:
